Final update probably

+++ Program.cs
@@ -9,13 +9,25 @@ using IniParser.Model;
 using Ookii.Dialogs.Wpf;
 using IWshRuntimeLibrary;
 using Mayerch1.GithubUpdateCheck;
+using Microsoft.Win32;
 
 
 
 static class Program
 {
     // #### Oculus Kill Switch Version ####
-    static readonly string AppVersion = "1.6.7";
+    static readonly string AppVersion = "1.6.8";
+
+    // #### Credit to @ndepoel ####
+    public static DirectoryInfo GetOculusBaseDirectory()
+    {
+        string OculusBaseVarName = @"OculusBase";
+        string oculusBasePath = Environment.GetEnvironmentVariable(OculusBaseVarName);
+        if (oculusBasePath == null)
+            return null;
+
+        return Directory.Exists(oculusBasePath) ? new DirectoryInfo(oculusBasePath) : null;
+    }
 
     static string GetMD5Hash(string filePath)
     {
@@ -84,21 +96,49 @@ static class Program
 
     }
 
-    static void MakeConfig()
+    static void CheckConfig()
     {
-        if (System.IO.File.Exists("OculusKillSwitch.ini") != true)
+        if (System.IO.File.Exists("OculusDash.exe") == false)
         {
-            System.IO.File.Create("OculusKillSwitch.ini").Close();
+            try
+            {
+                Process.Start("explorer.exe", GetOculusBaseDirectory().FullName + "Support\\oculus-dash\\dash\\bin");
+
+                DialogResult nuhuhbox = MessageBox.Show("I'm not in the right directory, I go here.\nAfter I close, would you please move me?", "Oculus Kill Switch", MessageBoxButtons.OK, MessageBoxIcon.Warning);
+                if (nuhuhbox == DialogResult.OK)
+                {
+                    Console.WriteLine("WRONG LOCATION/DASH NOT PRESENT");
+                    Application.Exit();
+                }
+            }
+            catch (Exception x)
+            {
+                Process.Start("explorer.exe", @"https://www.oculus.com/download_app/?id=1582076955407037");
+                DialogResult Dialog1 = MessageBox.Show("Looks like you've never installed the Oculus Software, please install Oculus Software before using this program, ya goof.", "Oculus Kill Switch", MessageBoxButtons.OK);
+                if (Dialog1 == DialogResult.OK)
+                {
+                    Console.WriteLine(x);
+                    Console.WriteLine("NOT INSTALLED AT ALL???");
+                    Application.Exit();
+                }
+            }
         }
-        var parser = new FileIniDataParser();
-        IniData configdata = parser.ReadFile("OculusKillSwitch.ini");
-        if (new FileInfo("OculusKillSwitch.ini").Length == 0)
-        {
-            configdata["OculusKillSwitch"]["IgnoreOculusClient"] = "false";
-            configdata["OculusKillSwitch"]["DontShowShortcutDialog"] = "false";
-            configdata["OculusKillSwitch"]["IgnoreUpdate"] = "false";
-            parser.WriteFile("OculusKillSwitch.ini", configdata);
+        else {
+            if (System.IO.File.Exists("OculusKillSwitch.ini") != true)
+            {
+                System.IO.File.Create("OculusKillSwitch.ini").Close();
+            }
+            var parser = new FileIniDataParser();
+            IniData configdata = parser.ReadFile("OculusKillSwitch.ini");
+            if (new FileInfo("OculusKillSwitch.ini").Length == 0)
+            {
+                configdata["OculusKillSwitch"]["IgnoreOculusClientOpen"] = "false";
+                configdata["OculusKillSwitch"]["DontShowShortcutDialog"] = "false";
+                configdata["OculusKillSwitch"]["IgnoreUpdate"] = "false";
+                parser.WriteFile("OculusKillSwitch.ini", configdata);
+            }
         }
+        
     }
 
     static void CheckForUpdate()
@@ -142,7 +182,7 @@ static class Program
 
     static void Main()
     {
-        MakeConfig();
+        CheckConfig();
         Application.SetHighDpiMode(HighDpiMode.PerMonitorV2);
         CheckForUpdate();
         Process[] DashCheck = Process.GetProcessesByName("OculusDash");
@@ -160,55 +200,67 @@ static class Program
             try
             {
                 activeFileHash = GetMD5Hash("OculusDash.exe");
+                configdata["OculusKillSwitch"]["OculusDashExecHash"] = activeFileHash;
+                parser.WriteFile("OculusKillSwitch.ini", configdata);
                 MakeShortcut();
+                try
+                {
+                    System.IO.File.Exists("OculusDash.exe.bak");
+                    backupFileHash = GetMD5Hash("OculusDash.exe.bak");
+                    
+                }
+                catch (Exception)
+                {
+                    if (activeFileHash == "9DB7CC8B646A01C60859B318F85E65D0")
+                    {
+                        MessageBox.Show("Oculus Dash backup couldn't be found. If it's in a different directory, move it here. If you don't have it, you may need to reinstall the Oculus app to get it back.", "Oculus Kill Switch", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        whoops = true;
+                        Console.WriteLine("NO BACKUP");
+                    }
+                    else if (activeFileHash == configdata["OculusKillSwitch"]["OculusDashExecHash"])
+                    {
+                        DialogResult Dialog1 = MessageBox.Show("The Oculus Killer isn't installed, or the backup file has been renamed. Would you like for me to download it for you?", "Oculus Kill Switch", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+                        if (Dialog1 == DialogResult.Yes)
+                        {
+                            using (WebClient client = new WebClient())
+                            {
+                                client.DownloadFile("https://github.com/kaitlyndotmoe/OculusKiller/releases/latest/download/OculusDash.exe", "OculusDash.exe.bak");
+                            }
+
+                        }
+                        else
+                        {
+                            whoops = true;
+                            Console.WriteLine("NO KILLER");
+                            Application.Exit();
+                        }
+                    }
+                    else
+                    {
+                        DialogResult Dialog1 = MessageBox.Show("Looks like the Oculus Client updated since this was last ran, give me a sec to update the stored hash value for the Oculus Client.", "Oculus Kill Switch", MessageBoxButtons.OK);
+                        if (Dialog1 == DialogResult.OK)
+                        {
+                            configdata["OculusKillSwitch"]["OculusDashExecHash"] = GetMD5Hash("OculusDash.exe");
+                            parser.WriteFile("OculusKillSwitch.ini", configdata);
+                        }
+                        else
+                        {
+                            whoops = true;
+                            Console.WriteLine("SHOULDA PRESSED OK");
+                            Application.Exit();
+                        }
+                    }
+                }
 
             }
             catch (Exception ex)
             {
+                activeFileHash = "null";
                 Console.WriteLine(ex);
                 whoops = true;
                 Console.WriteLine("WRONG LOCATION/DASH NOT PRESENT");
-                activeFileHash = "null";
-                Process.Start("explorer.exe", @"C:\Program Files\Oculus\Support\oculus-dash\dash\bin");
-                DialogResult nuhuhbox = MessageBox.Show("I'm not in the right directory, I go here.\nAfter I close, would you please move me?", "Oculus Kill Switch", MessageBoxButtons.OK, MessageBoxIcon.Warning);
-                if (nuhuhbox == DialogResult.OK)
-                {
-                    whoops = true;
-                    Console.WriteLine("WRONG LOCATION/DASH NOT PRESENT");
-                    Application.Exit();
-                }
             }
-            try
-            {
-                backupFileHash = GetMD5Hash("OculusDash.exe.bak");
-            }
-            catch (Exception)
-            {
-                if (activeFileHash == "9DB7CC8B646A01C60859B318F85E65D0")
-                {
-                    MessageBox.Show("Oculus Dash backup couldn't be found. If it's in a different directory, move it here. If you don't have it, you may need to reinstall the Oculus app to get it back.", "Oculus Kill Switch", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                    whoops = true;
-                    Console.WriteLine("NO BACKUP");
-                }
-                else if (activeFileHash == "DD65C6B1E5578AE812966798101F1379")
-                {
-                    DialogResult Dialog1 = MessageBox.Show("The Oculus Killer isn't installed, or the backup file has been renamed. Would you like for me to download it for you?", "Oculus Kill Switch", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
-                    if (Dialog1 == DialogResult.Yes)
-                    {
-                        using (WebClient client = new WebClient())
-                        {
-                            client.DownloadFile("https://github.com/kaitlyndotmoe/OculusKiller/releases/latest/download/OculusDash.exe", "OculusDash.exe.bak");
-                        }
-
-                    } else
-                    {
-                        whoops = true;
-                        Console.WriteLine("NO KILLER");
-                        Application.Exit();
-                    }
-                }
-            }
-            if (OculusAppCheck.Length > 0 == true && configdata["OculusKillSwitch"]["IgnoreOculusClient"] != "true" && whoops == false)
+            if (OculusAppCheck.Length > 0 == true && configdata["OculusKillSwitch"]["IgnoreOculusClientOpen"] != "true" && whoops == false)
             {
                 using (TaskDialog dialog = new TaskDialog())
                 {
@@ -224,7 +276,7 @@ static class Program
                     TaskDialogButton result = dialog.ShowDialog();
                     if (dialog.IsVerificationChecked)
                     {
-                        configdata["OculusKillSwitch"]["IgnoreOculusClient"] = "true";
+                        configdata["OculusKillSwitch"]["IgnoreOculusClientOpen"] = "true";
                         parser.WriteFile("OculusKillSwitch.ini", configdata);
                     }
                     if (result == butCancel)
@@ -261,17 +313,20 @@ static class Program
             // It'll be fiiiiiine
             if (Dialog0 == DialogResult.OK)
             {
-
-                System.IO.File.Move("OculusDash.exe.bak", "tempkill.exe");
-                System.IO.File.Move("OculusDash.exe", "OculusDash.exe.bak");
-                System.IO.File.Move("tempkill.exe", "OculusDash.exe");
-
                 if (activeFileHash != "9DB7CC8B646A01C60859B318F85E65D0")
                 {
+                    // Stock -> Default
+                    System.IO.File.Move("OculusDash.exe.bak", "tempkill.exe");
+                    System.IO.File.Move("OculusDash.exe", "OculusDash.exe.bak");
+                    System.IO.File.Move("tempkill.exe", "OculusDash.exe");
                     MessageBox.Show("Successfully enabled the Oculus Killer!", "Oculus Kill Switch", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 }
                 else
                 {
+                    // Killer -> Stock
+                    System.IO.File.Move("OculusDash.exe.bak", "tempbak.exe");
+                    System.IO.File.Move("OculusDash.exe", "OculusDash.exe.bak");
+                    System.IO.File.Move("tempbak.exe", "OculusDash.exe");
                     MessageBox.Show("Successfully disabled the Oculus Killer!", "Oculus Kill Switch", MessageBoxButtons.OK, MessageBoxIcon.Information);
                 }
             }
